test: split tests into unit and integration
